added caching system
