+++ server/api/states-map.js
@@ -2,6 +2,8 @@ const axios = require('axios');
 const generateXML = require('@mapbox/geojson-mapnikify');
 var mapnik = require('mapnik');
 var Jimp = require("jimp");
+const fs = require('fs');
+const path = require('path');
 
 const rangeSettings = {
     ranges: [
@@ -60,8 +62,8 @@ module.exports.statesMap =  async (req, res) => {
         theme = "light";
     }
     
-    const repsonse = await axios.get("https://opendata.arcgis.com/datasets/ef4b445a53c1406892257fe63129a8ea_0.geojson");
-    const geojson = repsonse.data;
+    const rawdata = fs.readFileSync(path.resolve(__dirname, '../cache/statesMap.json'));
+    const geojson = JSON.parse(rawdata);
 
     for (const feature of geojson.features) {
         feature.properties.fill = mapCasesToColor(feature.properties.faelle_100000_EW);
