+++ server/cache/states.json
@@ -1,0 +1,1 @@
+{"lastUpdate":1607727600000,"states":[{"name":"Schleswig-Holstein","code":"SH","count":17515,"weekIncidence":72.664082213038,"casesPer100k":603.180758275526,"deaths":290},{"name":"Hamburg","code":"HH","count":28620,"weekIncidence":108.431276062348,"casesPer100k":1549.32756909855,"deaths":462},{"name":"Niedersachsen","code":"NI","count":83935,"weekIncidence":86.0562589509018,"casesPer100k":1050.02647115045,"deaths":1403},{"name":"Bremen","code":"HB","count":11488,"weekIncidence":127.715420682852,"casesPer100k":1686.43075035012,"deaths":150},{"name":"Nordrhein-Westfalen","code":"NW","count":310932,"weekIncidence":158.932683784303,"casesPer100k":1732.47991987172,"deaths":4422},{"name":"Hessen","code":"HE","count":105934,"weekIncidence":172.866757420389,"casesPer100k":1684.67958422921,"deaths":1716},{"name":"Rheinland-Pfalz","code":"RP","count":54969,"weekIncidence":153.203434473167,"casesPer100k":1342.70401619188,"deaths":837},{"name":"Baden-Württemberg","code":"BW","count":182958,"weekIncidence":174.93072768408,"casesPer100k":1648.21176617695,"deaths":3353},{"name":"Bayern","code":"BY","count":253069,"weekIncidence":192.735290619538,"casesPer100k":1928.18339902735,"deaths":4812},{"name":"Saarland","code":"SL","count":15460,"weekIncidence":170.637570461461,"casesPer100k":1566.54206611294,"deaths":296},{"name":"Berlin","code":"BE","count":77208,"weekIncidence":170.132587871179,"casesPer100k":2104.05203337466,"deaths":790},{"name":"Brandenburg","code":"BB","count":25818,"weekIncidence":124.787213414685,"casesPer100k":1023.75477468711,"deaths":477},{"name":"Mecklenburg-Vorpommern","code":"MV","count":7870,"weekIncidence":78.1649336064442,"casesPer100k":489.385861163656,"deaths":96},{"name":"Sachsen","code":"SN","count":82226,"weekIncidence":342.094774250602,"casesPer100k":2019.31693521393,"deaths":1591},{"name":"Sachsen-Anhalt","code":"ST","count":17703,"weekIncidence":135.184268870439,"casesPer100k":806.594914665785,"deaths":263},{"name":"Thüringen","code":"TH","count":24811,"weekIncidence":209.714359105606,"casesPer100k":1162.99127487018,"deaths":508}]}
