+++ server/docker-compose.yaml
@@ -9,3 +9,5 @@ services:
       - DATABASE_USER
       - DATABASE_PASSWORD
       - DATABASE_NAME
+    volumes:
+      - ./cache:/usr/src/app/cache
