+++ server/server.js
@@ -1,12 +1,19 @@
 const express = require('express')
+const cors = require('cors');
 const { CronJob } = require('cron');
-const { updateDataset } = require('./cronjobs/updateDataset');
+
 const { general } = require('./api/general');
 const { states } = require('./api/states');
 const { statesMap } = require('./api/states-map');
 const { districts } = require('./api/districts');
 const { districtsMap } = require('./api/districts-map');
-const cors = require('cors')
+
+const { updateGeneral } = require('./cronjobs/updateGeneral');
+const { updateDistricts } = require('./cronjobs/updateDistricts');
+const { updateStates } = require('./cronjobs/updateStates');
+const { updateDistrictsMap } = require('./cronjobs/updateDistrictsMap');
+const { updateStatesMap } = require('./cronjobs/updateStatesMap');
+
 const app = express()
 const port = 3000
 
@@ -31,5 +38,19 @@ app.listen(port, () => {
   console.log(`Server running at http://localhost:${port}`)
 })
 
-var job = new CronJob('0 */10 * * * *', updateDataset);
+async function updateDataSources() {
+  try {
+    await updateGeneral();
+    await updateDistricts();
+    await updateStates();
+    await updateStatesMap();
+    await updateDistrictsMap();
+  } catch (error) {
+    console.log(error);
+  }
+}
+
+updateDataSources();
+
+var job = new CronJob('0 */15 * * * *', updateDataSources);
 job.start();
